Move things into integration tests
